+++ src/Thumbnails/index.types.ts
@@ -1,6 +1,9 @@
 import * as React from 'react'
 
-export type SpinoramaThumbnailsProps = {
+// Types
+import type { BoxProps } from '@mui/material'
+
+export type SpinoramaThumbnailsProps = BoxProps & {
 	/**
 	 * Determines how the slides are aligned.
 	 * Available options are:
@@ -19,16 +22,4 @@ export type SpinoramaThumbnailsProps = {
 	 * @default false
 	 */
 	reverse?: boolean
-
-	/**
-	 * The class name of the component.
-	 * Use this prop when you want to style the component with a className.
-	 */
-	className?: string
-
-	/**
-	 * Children elements to be rendered inside the component.
-	 * Use this prop when you want to pass components or elements as children to the component.
-	 */
-	children?: React.ReactNode
 }