+++ src/Wrapper/index.types.ts
@@ -1,6 +1,7 @@
-import * as React from 'react'
+// Types
+import type { BoxProps } from '@mui/material'
 
-export type SpinoramaWrapperProps = {
+export type SpinoramaWrapperProps = BoxProps & {
 	/**
 	 * Determines how the slides are aligned.
 	 * Available options are:
@@ -18,16 +19,4 @@ export type SpinoramaWrapperProps = {
 	 * @default false
 	 */
 	reverse?: boolean
-
-	/**
-	 * The class name of the component.
-	 * Use this prop when you want to style the component with a className.
-	 */
-	className?: string
-
-	/**
-	 * Children elements to be rendered inside the component.
-	 * Use this prop when you want to pass components or elements as children to the component.
-	 */
-	children?: React.ReactNode
 }